[qt] Extract ores.qt.headless: fix Windows build via real component split, not a workaround
Windows CI was failing with MSVC C4273 ("dllimport on a definition",
fatal under /WX) in HierarchyModelBuilder.cpp and OrgDocRenderer.cpp.
Root cause: these files define ORES_QT_API-decorated symbols but were
compiled a second time directly into ores.qt.api's test executable
(to dodge ores.qt.api.lib's Widgets/display requirement for headless
testing) — without ORES_QT_API_LIBRARY defined there, export.hpp
expands ORES_QT_API to dllimport for a TU that's also defining the
symbol, which MSVC flags as inconsistent. GCC/Clang don't warn on
this, which is why it went unnoticed until Windows CI actually ran
against it.

A one-line compile-definition fix would have silenced the specific
warning, but the underlying shape was the real problem: logic with
zero Qt6::Widgets dependency (OrgDocRenderer, TestScenarioResultsWriter,
EnvironmentMetadata, HierarchyModelBuilder, FontUtils, RepoFileFinder,
parse_trade_instrument/IInstrumentFormPopulator) was physically living
inside ores.qt.api, which as a whole does require Widgets — forcing
the test target to duplicate-compile specific .cpp files rather than
just link a proper headless-safe library.

Extracted all of it into a new ores.qt.headless component (own
export.hpp/ORES_QT_HEADLESS_API macro — deliberately distinct from
ores.qt.api's ORES_QT_API, since ores.qt.api consumes this component's
headers directly and a shared macro name would silently clash rather
than fail loudly). ores.qt.api.lib now links it PUBLIC (transitively
available to every domain plugin, unchanged include paths for existing
consumers). ores.qt.api's tests/ directory retired — every test file
in it was exactly the logic that moved; ores.qt.headless.tests is the
new home, now linking the real shared library instead of duplicate-
compiling source files, so this class of bug can't recur.

Verified: ores.qt.headless.lib/tests, ores.qt.api.lib, every ores.qt.*
plugin, and ores.qt.exe all build and link clean; full test suite
passes (198 assertions); the QA Validation Runner panel (the feature
that most exercises this code) verified live end-to-end via
--open-scenario after the extraction.
