+++ projects/CMakeLists.txt
@@ -104,6 +104,7 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.variability/service)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.telemetry/service)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.scheduler/service)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.synthetic/service)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.qt/headless)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.qt/api)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.qt/workspace)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ores.qt/admin)
