+++ projects/ores.qt/admin/src/AccountItemDelegate.cpp
@@ -22,7 +22,7 @@
 #include "ores.qt/ClientAccountModel.hpp"
 #include "ores.qt/ColorConstants.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include <QApplication>
 #include <QPainter>
 #include <QStyleOptionViewItem>
