+++ projects/ores.qt/admin/src/SystemSettingItemDelegate.cpp
@@ -20,7 +20,7 @@
 #include "ores.qt/SystemSettingItemDelegate.hpp"
 #include "ores.qt/ClientSystemSettingModel.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include <QApplication>
 #include <QPainter>
 #include <QStyleOptionViewItem>
