+++ projects/ores.qt/api/src/ClientResultItemDelegate.cpp
@@ -20,7 +20,7 @@
 #include "ores.qt/ClientResultItemDelegate.hpp"
 #include "ores.qt/ClientResultModel.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include <QApplication>
 #include <QPainter>
 #include <QStyleOptionViewItem>
