+++ projects/ores.qt/api/src/EntityItemDelegate.cpp
@@ -20,7 +20,7 @@
 #include "ores.qt/EntityItemDelegate.hpp"
 #include "ores.qt/ColorConstants.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include <QApplication>
 #include <QStyleOptionViewItem>
