+++ projects/ores.qt/application/src/EventViewerDialog.cpp
@@ -21,7 +21,7 @@
 #include "ores.qt/EventViewerDialog.hpp"
 #include "ores.platform/time/datetime.hpp"
 #include "ores.qt/ClientManager.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include "ores.qt/WidgetUtils.hpp"
 #include <QApplication>
 #include <QClipboard>
