+++ projects/ores.qt/application/src/main.cpp
@@ -19,7 +19,7 @@
  */
 #include "ores.logging/make_logger.hpp"
 #include "ores.qt/CommandLineParser.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include "ores.qt/IconUtils.hpp"
 #include "ores.qt/MainWindow.hpp"
 #include "ores.qt/PluginRegistry.hpp"
