+++ projects/ores.qt/compute/src/ServiceDashboardMdiWindow.cpp
@@ -23,7 +23,7 @@
 #include "ores.platform/time/datetime.hpp"
 #include "ores.qt/ColorConstants.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include "ores.qt/IconUtils.hpp"
 #include "ores.qt/UiPersistence.hpp"
 #include "ores.telemetry.core/messaging/service_samples_protocol.hpp"
