+++ projects/ores.qt/headless/CMakeLists.txt
@@ -17,3 +17,4 @@
 #
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/modeling)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/src)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/tests)
