+++ projects/ores.qt/headless/include/ores.qt.headless/EnvironmentMetadata.hpp
@@ -20,7 +20,7 @@
 #ifndef ORES_QT_ENVIRONMENT_METADATA_HPP
 #define ORES_QT_ENVIRONMENT_METADATA_HPP
 
-#include "ores.qt/export.hpp"
+#include "ores.qt.headless/export.hpp"
 #include <QString>
 
 namespace ores::qt {
@@ -48,7 +48,7 @@ struct environment_metadata final {
  * launcher (see =compass client start --instance-name=), empty if
  * unset.
  */
-ORES_QT_API environment_metadata capture_environment_metadata(const QString& reference_path = {});
+ORES_QT_HEADLESS_API environment_metadata capture_environment_metadata(const QString& reference_path = {});
 
 }
