+++ projects/ores.qt/headless/include/ores.qt.headless/IInstrumentFormPopulator.hpp
