+++ projects/ores.qt/headless/modeling/ores.qt.headless.puml
@@ -1,0 +1,39 @@
+' -*- mode: plantuml; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
+'
+' Copyright (C) 2026 Marco Craveiro <marco.craveiro@gmail.com>
+'
+' This program is free software; you can redistribute it and/or modify it under
+' the terms of the GNU General Public License as published by the Free Software
+' Foundation; either version 3 of the License, or (at your option) any later
+' version.
+'
+' This program is distributed in the hope that it will be useful, but WITHOUT
+' ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+' FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+'
+' You should have received a copy of the GNU General Public License along with
+' this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
+' Street, Fifth Floor, Boston, MA 02110-1301, USA.
+'
+@startuml
+
+title ores.qt.headless Component
+
+package "ores.qt.headless" {
+  [OrgDocRenderer]
+  [TestScenarioResultsWriter]
+  [EnvironmentMetadata]
+  [HierarchyModelBuilder]
+  [RepoFileFinder]
+  [FontUtils]
+  [parse_trade_instrument]
+  [IInstrumentFormPopulator]
+}
+
+[ores.qt.api] --> [ores.qt.headless] : links (PUBLIC)
+[ores.qt.api.tests] --> [ores.qt.headless] : links directly\n(no QApplication/display needed)
+[OrgDocRenderer] --> [ores.orgmode] : renders parsed docs
+[parse_trade_instrument] --> [ores.trading.api] : domain types
+[TestScenarioResultsWriter] --> [EnvironmentMetadata]
+
+@enduml
