+++ projects/ores.qt/headless/src/CMakeLists.txt
@@ -15,71 +15,37 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
 # Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
-set(name "ores.qt.api")
+set(name "ores.qt.headless")
 set(lib_binary_name ${name})
 set(lib_target_name ${name}.lib)
-set(ORES_QT_API_DIR ${CMAKE_CURRENT_SOURCE_DIR}/..)
-
-# Qt code generation (MOC for Q_OBJECT, UIC for .ui forms)
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTOUIC ON)
+set(ORES_QT_HEADLESS_DIR ${CMAKE_CURRENT_SOURCE_DIR}/..)
 
 file(GLOB_RECURSE files RELATIVE
     "${CMAKE_CURRENT_SOURCE_DIR}/"
     "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
 
-# Headers must be listed for AUTOMOC to find Q_OBJECT declarations.
-file(GLOB_RECURSE HEADERS "${ORES_QT_API_DIR}/include/*.hpp")
-# UI forms for AUTOUIC.
-file(GLOB FORMS "${ORES_QT_API_DIR}/ui/*.ui")
+add_library(${lib_target_name} SHARED ${files})
 
-add_library(${lib_target_name} SHARED ${files} ${HEADERS} ${FORMS})
-
-# Define ORES_QT_API_LIBRARY when building this target so export.hpp
+# Define ORES_QT_HEADLESS_LIBRARY when building this target so export.hpp
 # expands to BOOST_SYMBOL_EXPORT instead of BOOST_SYMBOL_IMPORT.
-target_compile_definitions(${lib_target_name} PRIVATE ORES_QT_API_LIBRARY)
+target_compile_definitions(${lib_target_name} PRIVATE ORES_QT_HEADLESS_LIBRARY)
 
 set_target_properties(${lib_target_name} PROPERTIES
     OUTPUT_NAME ${lib_binary_name}
-    AUTOUIC_SEARCH_PATHS ${ORES_QT_API_DIR}/ui
     VERSION   ${PROJECT_VERSION}
     SOVERSION ${PROJECT_VERSION_MAJOR})
 
-# Export the ui/ directory path so consumers (e.g. ores.qt) can add it to their
-# AUTOUIC_SEARCH_PATHS without hard-coding the source layout.
-set(ORES_QT_API_UI_DIR "${ORES_QT_API_DIR}/ui" CACHE INTERNAL "ores.qt.api UI forms directory")
-
 target_include_directories(${lib_target_name} PUBLIC
-    ${ORES_QT_API_DIR}/include)
-# The autogen include dir (ui_*.h) is only needed internally; keep it PRIVATE.
-target_include_directories(${lib_target_name} PRIVATE
-    "${CMAKE_CURRENT_BINARY_DIR}/${lib_target_name}_autogen/include")
+    ${ORES_QT_HEADLESS_DIR}/include)
 
 target_link_libraries(${lib_target_name}
     PUBLIC
-        ores.qt.headless.lib
-        ores.nats.lib
-        ores.eventing.api.lib
-        ores.iam.api.lib
-        ores.refdata.api.lib
-        ores.compute.api.lib
-        ores.scheduler.api.lib
-        ores.dq.api.lib
-        ores.http.api.lib
-        ores.assets.api.lib
         ores.trading.api.lib
-        ores.workspace.api.lib
-        ores.variability.api.lib
         ores.orgmode.lib
         ores.logging.lib
         ores.utility.lib
-        ores.platform.lib
         Qt6::Core
-        Qt6::Gui
-        Qt6::Widgets
-        Qt6::Concurrent
-        Qt6::Network
-        Qt6::Svg)
+        Qt6::Gui)
 
 install(TARGETS ${lib_target_name}
     LIBRARY DESTINATION lib
