+++ projects/ores.qt/headless/src/HierarchyModelBuilder.cpp
@@ -17,25 +17,38 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#ifndef ORES_QT_REPO_FILE_FINDER_HPP
-#define ORES_QT_REPO_FILE_FINDER_HPP
-
-#include "ores.qt/export.hpp"
+#include "ores.qt.headless/HierarchyModelBuilder.hpp"
+#include <QStandardItem>
+#include <QStandardItemModel>
 #include <QString>
+#include <boost/uuid/uuid_io.hpp>
 
 namespace ores::qt {
 
-/**
- * @brief Walk up from @p referencePath looking for @p filename at the
- * repo root (e.g. =.org-roam.db=, =compass.sh=). Empty if not found
- * within a handful of levels.
- *
- * Shared by anything that needs to locate a repo-root file from an
- * arbitrary doc/scenario path — e.g. =QaValidationRunnerWidget= and
- * =OrgDocViewerWindow= both need the org-roam index this way.
- */
-ORES_QT_API QString find_repo_file(const QString& referencePath, const QString& filename);
+namespace {
 
+QStandardItem* make_item(const ores::utility::domain::hierarchy_node& node) {
+    auto* item = new QStandardItem(QString::fromStdString(node.name));
+    item->setData(QString::fromStdString(boost::uuids::to_string(node.id)), Qt::UserRole);
+    item->setEditable(false);
+
+    for (const auto& child : node.children)
+        item->appendRow(make_item(child));
+
+    return item;
 }
 
-#endif
+} // namespace
+
+QStandardItemModel*
+HierarchyModelBuilder::build(const std::vector<ores::utility::domain::hierarchy_node>& roots) {
+    auto* model = new QStandardItemModel();
+    model->setHorizontalHeaderLabels({QStringLiteral("Name")});
+
+    for (const auto& root : roots)
+        model->appendRow(make_item(root));
+
+    return model;
+}
+
+}
