+++ projects/ores.qt/headless/src/parse_trade_instrument.cpp
@@ -17,9 +17,9 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#include "ores.qt/parse_trade_instrument.hpp"
+#include "ores.qt.headless/parse_trade_instrument.hpp"
 #include "ores.logging/make_logger.hpp"
-#include "ores.qt/IInstrumentFormPopulator.hpp"
+#include "ores.qt.headless/IInstrumentFormPopulator.hpp"
 #include "ores.utility/rfl/reflectors.hpp" // IWYU pragma: keep.
 #include "parse_swap_impl.hpp"
 #include <rfl/json.hpp>
