+++ projects/ores.qt/refdata/src/ChangeReasonItemDelegate.cpp
@@ -20,7 +20,7 @@
 #include "ores.qt/ChangeReasonItemDelegate.hpp"
 #include "ores.qt/ClientChangeReasonModel.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include <QApplication>
 #include <QPainter>
 #include <QStyleOptionViewItem>
