+++ projects/ores.qt/workflow/src/WorkflowMdiWindow.cpp
@@ -19,7 +19,7 @@
 #include "ores.qt/WorkflowMdiWindow.hpp"
 #include "ores.qt/ColorConstants.hpp"
 #include "ores.qt/DelegatePaintUtils.hpp"
-#include "ores.qt/FontUtils.hpp"
+#include "ores.qt.headless/FontUtils.hpp"
 #include "ores.qt/IconUtils.hpp"
 #include "ores.qt/UiPersistence.hpp"
 #include "ores.qt/WorkflowStepLogWidget.hpp"
